+++ src/common/interfaces/module-metadata.interface.ts
@@ -2,7 +2,7 @@ import { NestModule } from './nest-module.interface';
 import { Controller } from './controller.interface';
 
 export interface ModuleMetadata {
-    modules?: NestModule[];
+    modules?: NestModule[] | any[];
     components?: any[];
     controllers?: Controller[] | any[];
     exports?: any[];